add prometheus metrics for our http handler
We're migrating from statsd to prometheus and this is one step in it.
This also gives us the chance to correct how we export our API metrics
from having them rolled up by name to have them rolled up by label, as
promql (used by yardstick and Google Metrics Explorer) and prometheus
wants.

We also give up having the success rate's denominator be based on the
requests that are in flight and move to using only the ones returned and
then use the loadbalancer's view of the success rate plus the in-flight
requests to detect requests that are stuck.

(We also need to set up some more request timeouts, but that's not this
patch.)

Updates AUT-199